models were fixed

+++ src/SmartMealService.Domain/Models/OrderingItem.cs
@@ -3,5 +3,5 @@
 public sealed class OrderingItem
 {
     public required MenuItem MenuItem { get; init; }
-    public required float Quantity { get; init; }
+    public required double Quantity { get; init; }
 }
